% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_years.R
\name{bind_years}
\alias{bind_years}
\title{All Years in One}
\usage{
bind_years(data)
}
\arguments{
\item{data}{A list of dataframes. Each element is a year dataset, containing info on workers, firms and establishment.}
}
\value{
The raw dataset (a tibble).
}
\description{
This function joins all the years in a full QP. This file creates the RAW QP dataset.
}
