% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_promotions.R
\name{clean_promotions}
\alias{clean_promotions}
\title{Clean the promotions variable}
\usage{
clean_promotions(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble with cleaned QD data.
}
\description{
Add two variables.
The first is months_since_promotion. This tells how many months since the last promotion.
The second is the dummy "promoted", which tells us if the worker was promoted since last October.
}
