% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tenure.R
\name{clean_tenure}
\alias{clean_tenure}
\title{Clean tenure variable}
\usage{
clean_tenure(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble with the variable tenure.
}
\description{
There are two variables from which tenure can be retrived. admission date (\code{adm_date}) and \code{antig}.
\code{antig} comes in years, but we want to have it in months.
So I create the \code{tenure} variable,
which takes preference from adm_date and goes to
antig when there is no adm_date reported. \code{tenure} is in months.
}
