% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wages.R
\name{clean_wages}
\alias{clean_wages}
\title{Clean the wage variables}
\usage{
clean_wages(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble with cleaned wage variables
}
\description{
This fucntion does several alterations to the wage variables.
A reminder that base wage is paid in 14 months since 1982 (the whole QP sample)
https://poligrafo.sapo.pt/fact-check/direito-ao-13-o-mes-em-portugal-teve-origem-numa-greve-geral-em-julho-de-1962
\itemize{
\item Detects maternities and baixas
\item Binds diuturnidades in regular_benefits
\item A total wage that considerest the 14 months.
\item log real hourly wage that contains all wages
\item fte is a full time equivalence just with base wages. The goal is to make it comparable with the MW. It is a base wage normalized for the normal hours of that time. Which is 44 until 96 and 40h after.
}
}
