% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvrt_date_sas.R
\name{cnvrt_date_sas}
\alias{cnvrt_date_sas}
\title{Convert date from sas to R}
\usage{
cnvrt_date_sas(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A date in a format \%Y-\%m-\%d
}
\description{
This function converst the numeric sas dates into dates.
Sas counts the days since 1582-10-14.
}
