% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_educ.R
\name{crosswalk_educ}
\alias{crosswalk_educ}
\title{Education Crosswalk for QP}
\usage{
crosswalk_educ(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
A varaible "educ", consistent thorugh the years and with a mode applied.
}
\description{
This function applies a crosswalk to the education variable and also a mode.
The mode means that we do not allow education to vary within each worker.
I apply this mode for several the following reason:
Firstly, only for a small fraction does education vary through time.
Secondly, from the ones that have chages, there are a lot of decreases in education level.
I'm aware that the code of Marta Silva has an option to let the education level
increase, but never decrease.
However, I think that given the noise of this variable means that
an increases are also a reflection of errors.

I had the help of Francisco Weinholtz for this function.

The levels of the educ variable:\tabular{ll}{
   Level \tab Classification \cr
   1 \tab Less than primary \cr
   2 \tab Primary \cr
   3 \tab Basic (6 years) \cr
   4 \tab Basic (9 years) \cr
   5 \tab High School \cr
   6 \tab Polytechnic \cr
   7 \tab College \cr
   9 \tab Not defined \cr
}


In the examples there is a code to transform the variable into these names.
}
\examples{
# educ_label =
#   dplyr::case_when(
#     educ == 1 ~ "Less than primary",
#     educ == 2 ~ "Primary",
#     educ == 3 ~ "Basic (6 years)",
#     educ == 4 ~ "Basic (9 years)",
#     educ == 5 ~ "High School",
#     educ == 6 ~ "Polytechnic",
#     educ == 7 ~ "College",
#     educ == 9 ~ "Not defined") \%>\%
#   as_factor() \%>\%
#   # to determine the order of the factors
#   fct_relevel("Less than primary","Primary","Basic (6 years)",
#               "Basic (9 years)","High School","Polytechnic",
#               "College", "Not defined"))

}
