% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_firm.R
\name{crosswalk_firm}
\alias{crosswalk_firm}
\title{Crosswalk of the Firm ID, NPC_FIC}
\usage{
crosswalk_firm(data)
}
\arguments{
\item{data}{A tibble with firm inconsistencies.}
}
\value{
A tibble with the variable firm cleaned.
}
\description{
There are several problems with the firm ID that this function deals with.
The \code{firm} identifier comes from the variable \code{NPC_FIC}. And there are 2 problems:
(1) GEP allows for \code{NPC_FIC} to be recycled after 2010. If a firm had closed before 2010, its NPC_FIC could be reused;
(2) \code{NPC_FIC} is not valid/existent for some observations.
To solve it, I select either \code{nuemp} or \code{emp_id} for this firms, which are older firm IDs.
A valid firm ID (\code{NPC_FIC}) has 9 digits and starts with a 5.
To solve Problem 1, I build an algorithm that:
\enumerate{
\item Get all the firms that exist from 2011 to 2020;
\item Keep only the ones that did not exist in 2010;
\item Keep only the ones that existed before 2010;
\item Give a new ID to this firms, which is $NPC\_FIC \times 100$ (adding 2 0's).
}
}
