% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_occup.R
\name{crosswalk_occup}
\alias{crosswalk_occup}
\title{Occupation Crosswalk}
\usage{
crosswalk_occup(
  data,
  year_column,
  original_occup_3d,
  crosswalk_80_94 = data("occup_crosswalk_80_94"),
  crosswalk_10_94 = data("occup_crosswalk_10_94")
)
}
\arguments{
\item{data}{A tibble with an assembled QP.}

\item{year_column}{Not a string. Name of the year column in data.}

\item{original_occup_3d}{Not a string. Name of the column with occupation 3 digits column in data.}

\item{crosswalk_80_94}{Crosswalk from 1980 to 1994. A dataframe.}

\item{crosswalk_10_94}{Crosswalk from 2010 to 1994. A dataframe.}
}
\value{
A dataset with the "occup" column, which is crosswalked to the 1994 classification.
}
\description{
Do a crosswalk from the occupation classification of 2010 and 1980 both to
the classification of 1994.
There are 3 occupation classifications in the pediod of the full QP:\tabular{ll}{
   Name \tab Years \cr
   CNP 80 \tab (1980-1994) \cr
   CNP 94 \tab (1995-2009) \cr
   CPP 2010 \tab (2010-today) \cr
}
}
