% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_sector.R
\name{crosswalk_sector}
\alias{crosswalk_sector}
\title{Sector Crosswalk for QP}
\usage{
crosswalk_sector(data, year, original_3d_cae, new_variable = sector)
}
\arguments{
\item{data}{A tibble.}

\item{year}{Not a string. Name of the year column in data.}

\item{original_3d_cae}{Not a string. Name of the column with the original 3 digit CAE.}

\item{new_variable}{Name of the new variable. Default is "sector".}
}
\value{
A tibble with the column "sector", consistent throughout the years.
}
\description{
A sector crosswalk of sector at the 2 digit level. Creates the column "sector".

Crosswalk based on Pedro Raposo's files.

The classification is the following:\tabular{ll}{
   Code \tab Industry \cr
   1 \tab Agriculture, Animal Production, Hunt e Silvicultura \cr
   2 \tab Pesca \cr
   3 \tab Extracção de Produtos Energéticos \cr
   4 \tab "Indústrias Extractivas com Excepção da Extracção de Produtos Energéticos" \cr
   5 \tab "Indústrias Alimentares, das Bebidas e do Tabaco" \cr
   6 \tab "Indústria Têxtil" \cr
   7 \tab "Indústria do Couro e dos Produtos de Couro" \cr
   8 \tab "Indústrias da Madeira e da Cortiça e suas Obras" \cr
   9 \tab "Indústrias de Pasta de Papel e Cartão e seus Artigos, Edição e Impressão" \cr
   10 \tab "Fabricação de Coque, Produtos Petrolíferos Refinados e Combustível Nuclear" \cr
   11 \tab "Fabricação de Produtos Químicos e de Fibras Sintéticas ou Artificiais" \cr
   12 \tab "Fabricação de Artigos de Borracha e de Matérias Plásticas" \cr
   13 \tab "Fabricação de Outros Produtos Minerais Não Metálicos" \cr
   14 \tab "Indústrias Metalúrgicas de Base e de Produtos Metálicos" \cr
   15 \tab "Fabricação de Máquinas e de Equipamentos N.E." \cr
   16 \tab "Fabricação de Equipamento Eléctrico e de Óptica" \cr
   17 \tab "Fabricação de Material de Transporte" \cr
   18 \tab "Indústrias Transformadoras N.E." \cr
   19 \tab "Produção e Distribuição de Electricidade, de Gás e de Água" \cr
   20 \tab "Construção" \cr
   21 \tab "Comércio por Grosso e a Retalho; Reparação de Veículos e de Bens de Uso Pessoal" \cr
   22 \tab "Alojamento e Restauração (Restaurantes e Similares)" \cr
   23 \tab "Transportes, Armazenagem e Comunicações" \cr
   24 \tab "Actividades Financeiras" \cr
   25 \tab "Actividades Imobiliárias, Alugueres e Serviços Prestados às Empresas" \cr
   26 \tab "Administração Publica, Defesa e Segurança Social Obrigatória" \cr
   27 \tab "Educação" \cr
   28 \tab "Saúde e Acção Social" \cr
   29 \tab "Outras Actividades de Serviços Colectivos, Sociais e Pessoais" \cr
   30 \tab "Famílias com Empregados Domésticos" \cr
   31 \tab "Organismos Internacionais e Outras Instituições Extra-territoriais" \cr
}


Sector/Industry CAE Revisions:\tabular{ll}{
   Name \tab Years \cr
   Rev 1 \tab (1982-1994) \cr
   Rev 2 \tab (1995-2002) \cr
   Rev 2.1 \tab (2003-2006) \cr
   Rev 3 \tab 2007-today \cr
}


To the 2 digits agregation level, there is no difference between Rev 2 and Rev 2.1.
}
