% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_worker_duplicates.R
\name{delete_worker_duplicates}
\alias{delete_worker_duplicates}
\title{Delete Worker Duplicates}
\usage{
delete_worker_duplicates(data, type = "continuity")
}
\arguments{
\item{data}{A tibble.}

\item{type}{A string determining the type of selection. "none" keeps the duplicates. "continuity" uses the algorithm of Salema (2022) to select. "highest" uses the highest wage and hours to select the duplicate. "all" deletes all worker duplicates.}
}
\value{
A tibble with duplicates selected. If type is set to "continuity" or "highest", there is also the column "w_g", indicating if that observation would have been eliminated using "all".
}
\description{
A worker can hold several jobs simultaneously.
In that case, workers will have more than one observation per year, reported by different firms.
}
