% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm_names_standard.R
\name{firm_names_standard}
\alias{firm_names_standard}
\title{Rename and Reorder Firm Info}
\usage{
firm_names_standard(data)
}
\arguments{
\item{data}{A list with datasets. Each element of the list is a dataset.}
}
\value{
A list with firm level data with consistent columns.
}
\description{
This function selects and renames the variables of firm-level data.
There are inconsistencies in the years. Same variables, different names.
But also new variables.
}
