% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_FirmEst.R
\name{join_FirmEst}
\alias{join_FirmEst}
\title{Join Firms to Establishments}
\usage{
join_FirmEst(data_firm, data_est)
}
\arguments{
\item{data_firm}{A list with a dataset for each year; containing firm-level info.}

\item{data_est}{A list with a dataset for each year; containing establishment-level info.}
}
\value{
A list with joined firm and establishment data.
}
\description{
Connect firms to their establishments, year by year, maintaining the list format.
From 1985 to 2009, firms and establishments are joined using
"nuemp" (and NPC_FIC remains in the data). From 2010 onwards,
I use NPC_FIC for the join.
}
