% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_Worker.R
\name{join_Worker}
\alias{join_Worker}
\title{Join Worker Level Info}
\usage{
join_Worker(data_firm_est, data_worker)
}
\arguments{
\item{data_firm_est}{A list with a dataset for each year; containing firm-establishment-level info.}

\item{data_worker}{A list with a dataset for each year; containing worker-level info.}
}
\value{
A list with joined worker, firm and establishment data.
}
\description{
Connect workers to their establishments and firms, year by year, maintaining the list format.
This assumes that there is already a list with joined firm and establishment info.
From 1985 to 2009, workers are joined to firms using "nuemp" and to
establishments using "nuest". From 2010 onwards,
I use "NPC_FIC" and "estab_id", respectively.
}
