% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_assemble.R
\name{qp_assemble}
\alias{qp_assemble}
\title{QP, Assemble!}
\usage{
qp_assemble(path_worker_dir, path_firm_dir, path_est_dir)
}
\arguments{
\item{path_worker_dir}{Path to the directory with worker-level data. Format of the files: QP_Trab_yyyy}

\item{path_firm_dir}{Path to the directory with firm-level data. Format of the files: QP_Emp_yyyy}

\item{path_est_dir}{Path to the directory with establishment-level data. Format of the files: QP_Est_yyyy}
}
\value{
A QP raw dataset, assembled (all years in one panel).
}
\description{
This function takes separated files of QP, with information divided by year and within
each year divided by workers, firms and establishments; and bind them all into one dataset.
To use the function, make sure all dataframes have coherent names and are in the same directory.
}
\examples{
# set the parallel processing

}
