% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_clean_vars.R
\name{qp_clean_vars}
\alias{qp_clean_vars}
\title{Clean several variables in QP}
\usage{
qp_clean_vars(data)
}
\description{
This function cleans the following varaibles:
\itemize{
\item male
\item age
\item nationality (work in progress)
\item wages
}
}
