% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_crosswalks.R
\name{qp_crosswalks}
\alias{qp_crosswalks}
\title{Varaibles Consistency}
\usage{
qp_crosswalks(data)
}
\arguments{
\item{data}{A raw QP dataset (a tibble).}
}
\value{
description
}
\description{
Things that are done, include the following the changes to the variables:
(1) worker, delete NAs and put all NA's in 1985;
(2) cae_3;
(3) year founded;
(4) nut_2_firm;
(5) nut_2_est;
(6) firm;

The function is commented with each of the problems.
}
