% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_lags.R
\name{qp_lags}
\alias{qp_lags}
\title{Create Firm and Worker lag/lead variables}
\usage{
qp_lags(data)
}
\arguments{
\item{data}{A QP assembled sample as a tibble.}
}
\value{
A tibble with lag variables.
}
\description{
The goal of creating lag and lead variables is to avoid classifying them
after the restrictions. That is, if the goal is to know if a worker
separated, we should do it before the elimination of maternity,
for example.
}
