% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_names_standard.R
\name{worker_names_standard}
\alias{worker_names_standard}
\title{Rename and Reorder Worker Info}
\usage{
worker_names_standard(data)
}
\arguments{
\item{data}{A list with datasets. Each element of the list is a dataset.}
}
\value{
A list with worker level data with consistent columns.
}
\description{
This function selects and renames the variables of worker-level data.
There are inconsistencies in the years. Same variables, different names.
But also new variables.
}
